pyproject.toml
[tool.poetry]
authors = ["7rikaz_h785 <7rikaz.h785.stat2ltas41lcijad@gmail.com>"]
description = "This project is based on the original code of the [inteoryx / twitter-video-dl](https://github.com/inteoryx/twitter-video-dl) project, which allows users to download Twitter videos as MP4 files using Python, FFmpeg and URLs without the need for API keys. I forked this project for use in iOS Shortcuts application."
license = "Unlicense"
name = "twitter-video-dl-for-sc"
readme = "README.md"
version = "0.2.8"

[tool.poetry.dependencies]
python = "^3.11.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.292"
taskipy = "^1.12.0"
ffmpeg-python = "^0.2.0"
pre-commit = "^3.4.0"
tomlkit = "^0.12.1"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
pytest-html = "^4.1.1"

[tool.taskipy.tasks]
# poetry run task [task]
blackfix = "poetry run black twitter-video-dl-for-sc.py twitter-video-dl.py src tests ci"
ruffcheck = "poetry run ruff twitter-video-dl-for-sc.py twitter-video-dl.py src tests ci"
rufffix = "poetry run ruff twitter-video-dl-for-sc.py twitter-video-dl.py src tests ci --fix"

[tool.black]
target-version = ['py310']

[tool.ruff]
# Check Rules
select = ["E", "F", "I"]
# Ignore Error
ignore = ["E402","E501"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Assume Python 3.10.
target-version = "py310"

[build-system]
requires = [
    "setuptools",
    "requests>2.27.0"
]
build-backend = "setuptools.build_meta"
